fix: PageNotFoundView page and routing; add data loading in FrontLayoutView
